collision detection and CCD: Not exactly fixed but it works good enough
Former-commit-id: 3feb9b1156fc941390b4efe4c1540ed4c5dbd109
Former-commit-id: 8c97b0ade765731800ef09b722e0742e44443338

+++ src/org/dyn4j/geometry/Vector2.kt
@@ -66,10 +66,10 @@ import org.dyn4j.Epsilon
 class Vector2 {
 
     /** The magnitude of the x component of this [Vector2]  */
-    var x: Double = 0.toDouble()
+    var x: Double = 0.0
 
     /** The magnitude of the y component of this [Vector2]  */
-    var y: Double = 0.toDouble()
+    var y: Double = 0.0
 
     /** Default constructor.  */
     constructor() {
